add docs refactor configure step
