+++ src/espIdf/unitTest/testExecution.ts
@@ -16,8 +16,7 @@
  * limitations under the License.
  */
 
-import { copy, pathExists, readFile, writeFile } from "fs-extra";
-import { join } from "path";
+import { readFile } from "fs-extra";
 import {
   CancellationToken,
   ShellExecution,
@@ -27,148 +26,13 @@ import {
   TaskRevealKind,
   TaskScope,
   Uri,
-  extensions,
 } from "vscode";
-import { ESP } from "../../config";
 import { readParameter } from "../../idfConfiguration";
-import { appendIdfAndToolsToPath, startPythonReqsProcess } from "../../utils";
-import { buildCommand } from "../../build/buildCmd";
-import { verifyCanFlash } from "../../flash/flashCmd";
-import { flashCommand } from "../../flash/uartFlash";
-import { jtagFlashCommand } from "../../flash/jtagCmd";
+import { appendIdfAndToolsToPath } from "../../utils";
 import { OutputChannel } from "../../logger/outputChannel";
 import { parseStringPromise } from "xml2js";
 import { TaskManager } from "../../taskManager";
-
-export async function copyTestAppProject(
-  workspaceFolder: Uri,
-  testComponents: string[]
-) {
-  let unityAppDir: string = join(
-    extensions.getExtension(ESP.extensionID).extensionPath,
-    "templates",
-    "unity-app"
-  );
-  let destUnityAppDir = Uri.joinPath(workspaceFolder, "unity-app");
-  await copy(unityAppDir, destUnityAppDir.fsPath);
-  await updateTestComponents(destUnityAppDir, testComponents);
-  return destUnityAppDir;
-}
-
-export async function updateTestComponents(
-  unityApp: Uri,
-  testComponents: string[]
-) {
-  const cmakeListFile = Uri.joinPath(unityApp, "CMakeLists.txt");
-  if (pathExists(cmakeListFile.fsPath)) {
-    let content = await readFile(cmakeListFile.fsPath, "utf-8");
-    const projectMatches = content.match(/(project\(.*?\))/g);
-    if (projectMatches && projectMatches.length) {
-      content = content.replace(
-        /set\(TEST_COMPONENTS ""\)/g,
-        `set(TEST_COMPONENTS "${testComponents.join(" ")}")`
-      );
-      await writeFile(cmakeListFile.fsPath, content);
-    }
-  }
-}
-
-export async function checkPytestRequirements(workspaceFolder: Uri) {
-  const idfPath = readParameter("idf.espIdfPath", workspaceFolder);
-  const pythonBinPath = readParameter("idf.pythonBinPath", workspaceFolder);
-  const requirementsPath = join(
-    idfPath,
-    "tools",
-    "requirements",
-    "requirements.pytest.txt"
-  );
-  let checkResult: string;
-  try {
-    checkResult = await startPythonReqsProcess(
-      pythonBinPath,
-      idfPath,
-      requirementsPath
-    );
-  } catch (error) {
-    checkResult = error && error.message ? error.message : " are not satisfied";
-  }
-  if (checkResult.indexOf("are satisfied") > -1) {
-    return true;
-  }
-  return false;
-}
-
-export async function installPyTestPackages(
-  workspaceFolder: Uri,
-  cancelToken?: CancellationToken
-) {
-  const idfPath = readParameter("idf.espIdfPath", workspaceFolder);
-  const pythonBinPath = readParameter("idf.pythonBinPath", workspaceFolder);
-  const requirementsPath = join(
-    idfPath,
-    "tools",
-    "requirements",
-    "requirements.pytest.txt"
-  );
-
-  await runTaskForCommand(
-    workspaceFolder,
-    `"${pythonBinPath}" -m pip install --upgrade --no-warn-script-location -r "${requirementsPath}"`,
-    "Install Pytest",
-    cancelToken
-  );
-}
-
-export async function buildFlashTestApp(
-  workspaceFolder: Uri,
-  cancelToken: CancellationToken
-) {
-  const flashType = readParameter("idf.flashType", workspaceFolder);
-  let canContinue = await buildCommand(workspaceFolder, cancelToken, flashType);
-  if (!canContinue) {
-    return;
-  }
-  const port = readParameter("idf.port", workspaceFolder);
-  const flashBaudRate = readParameter("idf.flashBaudRate", workspaceFolder);
-  const idfPathDir = readParameter("idf.espIdfPath", workspaceFolder) as string;
-  const canFlash = await verifyCanFlash(flashBaudRate, port, workspaceFolder);
-  if (!canFlash) {
-    return;
-  }
-  if (flashType === ESP.FlashType.JTAG) {
-    canContinue = await jtagFlashCommand(workspaceFolder);
-  } else {
-    canContinue = await flashCommand(
-      cancelToken,
-      flashBaudRate,
-      idfPathDir,
-      port,
-      workspaceFolder,
-      flashType,
-      false
-    );
-  }
-  if (!canContinue) {
-    return;
-  }
-}
-
-export async function configurePyTestUnitApp(
-  workspaceFolder: Uri,
-  testComponents: string[],
-  cancelToken?: CancellationToken
-) {
-  const isPyTestInstalled = await checkPytestRequirements(workspaceFolder);
-  if (!isPyTestInstalled) {
-    await installPyTestPackages(workspaceFolder, cancelToken);
-  }
-  const unityTestApp = await copyTestAppProject(
-    workspaceFolder,
-    testComponents
-  );
-  await buildFlashTestApp(unityTestApp, cancelToken);
-  return unityTestApp;
-}
+import { Logger } from "../../logger/logger";
 
 export async function runPyTestWithTestCase(
   workspaceFolder: Uri,
@@ -203,10 +67,13 @@ export async function runPyTestWithTestCase(
         }
       }
     }
-
-    console.log(xmlResults);
   } catch (error) {
-    OutputChannel.appendLine(error);
+    const msg =
+      error && error.message
+        ? error.message
+        : "Error configuring PyTest Unit App for project";
+    OutputChannel.appendLine(msg, "idf-unit-test");
+    Logger.error(msg, error);
   }
 }
